title subcommands, preprocessor changes, etc
This is a big change that has some experimental work in it as well as
the implementation of the title subcommand.

For things outside of what could be considered libnrt, the header
defines remove the NRT_ prefix so that they don't get confused for the
internal nrt ones. This was causing an issue with the nrt/util.h and the
top-level util.h files where only the first was getting included and
causing some grief.

also, this adds a SUBCOMMAND() macro which simplifies the calling of a
subcommand by first calling the parse function, then passing the result
to the subcommand handler itself, reducing the amount of typing that has
to be done to call a subcommand.

also, the PATH_LENGTH constant has been moved into the top-level util.h
file, since it's something that might be used across the board.